requirements.txt
networkx==3.2.1
numpy==1.26.4
scipy==1.13.1
matplotlib==3.8.4
pandas==2.2.2
pillow==10.3.0
opencv-python==4.10.0.84
streamlit==1.36.0
pyvis==0.3.2
scikit-learn==1.5.1
jupyter==1.0.0
numba==0.60.0